\name{agromet-package}
\alias{agromet-package}
\alias{agromet}
\docType{package}

\title{AgroMet: tools for agro-meteorological analysis}
\description{
  AgroMet is a GUI (Graphical User Interface) to access function for loading,
  plotting, and saving maps and georeferenced data. It allows also to process
  such data, as interpolating a variable on a map using the AURELHY method.
}

\details{
  \tabular{ll}{
    Package: \tab agromet\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2010-07-11\cr
    License: \tab GPL 2 or above at your convenience.\cr
  }
}

\author{
  Philippe Grosjean, Numerical Ecology of Aquatic Systems,
  Mons University, Belgium.

  Programmed in the framework of an FAO project.

  Maintainer: Philippe Grosjean <Philippe.Grosjean@umons.ac.be>
}

\keyword{package}